requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.1.13
django-ckeditor==6.1.0
django-classy-tags==2.0.0
django-filer==2.0.2
django-js-asset==1.2.2
django-mptt==0.13.3
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-treebeard==4.4
easy-thumbnails==2.7.1
fontawesome-free==5.15.4
html5lib==1.1
Pillow==8.3.2
pycodestyle==2.7.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Unidecode==1.1.2
Unipath==1.1
webencodings==0.5.1
mysqlclient